requirements.txt
requests==2.23.0
python-archive==0.2
PySimpleGUI==4.24.0